+++ skosprovider_rdf/providers.py
@@ -12,7 +12,6 @@ from rdflib.term import Literal, URIRef
 from skosprovider_rdf.utils import text_, _df_writexml
 
 log = logging.getLogger(__name__)
-logging.basicConfig(level=logging.DEBUG)
 
 from skosprovider.providers import MemoryProvider
 from skosprovider.uri import (
@@ -37,6 +36,13 @@ DocumentFragment.writexml = _df_writexml
 
 
 class RDFProvider(MemoryProvider):
+
+    '''
+    Should the provider only take concepts into account explicitly linked 
+    to the conceptscheme?
+    '''
+    check_in_scheme = False
+
     '''
     A simple vocabulary provider that use an :class:`rdflib.graph.Graph`
     as input. The provider expects a RDF graph with elements that represent
@@ -47,12 +53,15 @@ class RDFProvider(MemoryProvider):
 
     def __init__(self, metadata, graph, **kwargs):
         self.graph = graph
+        self.check_in_scheme = False
         if not 'concept_scheme' in kwargs:
-            kwargs['concept_scheme'] = self._cs_from_graph(metadata)
+            kwargs['concept_scheme'] = self._cs_from_graph(metadata, **kwargs)
+        else:
+            self.check_in_scheme = True
         super(RDFProvider, self).__init__(metadata, [], **kwargs)
         self.list = self._from_graph()
 
-    def _cs_from_graph(self, metadata):
+    def _cs_from_graph(self, metadata, **kwargs):
         cslist = []
         for sub in self.graph.subjects(RDF.type, SKOS.ConceptScheme):
             uri = self.to_text(sub)
@@ -73,13 +82,30 @@ class RDFProvider(MemoryProvider):
         elif len(cslist) == 1:
             return cslist[0]
         else:
-            raise RuntimeError(
-                'This RDF file contains more than one ConceptScheme.'
-            )
+            if not 'concept_scheme_uri' in kwargs:
+                raise RuntimeError(
+                    'This RDF file contains more than one ConceptScheme. \
+                    Please specify one. The following schemes were found: \
+                    %s' % (", ".join([str(cs.uri) for cs in cslist]))
+                )
+            else:
+                self.check_in_scheme = True
+                csuri = kwargs['concept_scheme_uri']
+                filteredcslist = [cs for cs in cslist if cs.uri == csuri]
+                if len(filteredcslist) == 0:
+                    raise RuntimeError(
+                        'This RDF file contains more than one ConceptScheme. \
+                        You specified an unexisting one. The following schemes \
+                        were found: %s' % (", ".join([str(cs.uri) for cs in cslist]))
+                    )
+                else:
+                    return filteredcslist[0]
 
     def _from_graph(self):
         clist = []
         for sub, pred, obj in self.graph.triples((None, RDF.type, SKOS.Concept)):
+            if self.check_in_scheme and self._get_in_scheme(sub) != self.concept_scheme.uri:
+                    continue
             uri = self.to_text(sub)
             matches = {}
             for k in Concept.matchtypes:
@@ -101,6 +127,8 @@ class RDFProvider(MemoryProvider):
             clist.append(con)
 
         for sub, pred, obj in self.graph.triples((None, RDF.type, SKOS.Collection)):
+            if self.check_in_scheme and self._get_in_scheme(sub) != self.concept_scheme.uri:
+                    continue
             uri = self.to_text(sub)
             col = Collection(
                 id=self._get_id_for_subject(sub, uri), 
@@ -116,6 +144,20 @@ class RDFProvider(MemoryProvider):
             clist.append(col)
         self._fill_member_of(clist)
         return clist
+
+    def _get_in_scheme(self, subject):
+        '''
+        Determine if a subject is part of a scheme.
+
+        :param subject: Subject to get the sources for.
+        :returns: A URI for the scheme a subject is part of or None if
+            it's not part of a scheme.
+        '''
+        scheme = None
+        scheme = self.graph.value(subject, SKOS.inScheme)
+        if not scheme:
+            scheme = self.graph.value(subject, SKOS.topConceptOf)
+        return self.to_text(scheme) if scheme else None
 
     def _fill_member_of(self, clist):
         collections = list(set([c for c in clist if isinstance(c, Collection)]))